requirements/base.txt
# Python
# ------------------------------------------------------------------------------
pytz==2022.4
python-slugify==6.1.2
Pillow==9.2.0
argon2-cffi==21.3.0
redis==4.3.4
hiredis==2.0.0
twilio==7.14.2
timeago==1.0.16
psutil==5.9.2
bcrypt==4.0.0
dacite==1.6.0
awsebcli==3.20.3
beautifulsoup4==4.11.1

# Django
# ------------------------------------------------------------------------------
django==3.2.16
django-environ==0.9.0
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-redis==5.2.0
django-compressor==4.1
django-health-check==3.17.0
django-crontab==0.7.1
django-imagekit==4.1.0
django-seed==0.3.1
django-admin-numeric-filter==0.1.6

# Django Settings
# ------------------------------------------------------------------------------
django-cors-headers==3.13.0
drf-pretty-exception-handler==0.1.3
django-health-check==3.17.0


# Django Admin
# ------------------------------------------------------------------------------
https://github.com/leoyunhyung/django-jet/archive/dev.zip
django-reorder-admin==0.3.1
django-admin-relation-links==0.2.5
django-nested-admin==3.4.0
django-import-export==2.9.0
django-inline-actions==2.4.0
django-admin-rangefilter==0.9.0
django-advanced-filters==2.0.0
django-reverse-admin==2.9.6
django-nested-inline==0.4.5
django-summernote==0.8.20.0


# Django Models
# ------------------------------------------------------------------------------
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
phonenumberslite==8.12.56
django-annoying==0.10.6


# Django Views
# ------------------------------------------------------------------------------
django-filter==22.1
django-url-filter==0.3.15
django-extra-fields==3.0.2
drf-yasg==1.21.4


# DRF
# ------------------------------------------------------------------------------
djangorestframework==3.14.0
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.2.2
